test(api): Split tests into two different folders: integration, e2e
