+++ api/package.json
@@ -12,7 +12,7 @@
     "start:dev": "nest start --watch",
     "start:debug": "nest start --debug --watch",
     "lint": "eslint \"src/**/*.ts\" --fix",
-    "test:e2e": "jest --config ./test/jest-config.json -i",
+    "test": "jest --config ./test/jest-config.json -i",
     "transform:sections": "ts-node data/transform-sections.ts"
   },
   "dependencies": {
